Temporarily adapting the Janus model to the Orange Pi

+++ mindnlp/transformers/cache_utils.py
@@ -375,7 +375,7 @@ class DynamicCache(Cache):
             self.key_cache[layer_idx] = key_states
             self.value_cache[layer_idx] = value_states
         else:
-            self.key_cache[layer_idx] = ops.cat([self.key_cache[layer_idx], key_states], dim=-2)
+            self.key_cache[layer_idx] = ops.cat([self.key_cache[layer_idx], key_states.astype(mindspore.float32)], dim=-2)
             self.value_cache[layer_idx] = ops.cat([self.value_cache[layer_idx], value_states], dim=-2)
 
         return self.key_cache[layer_idx], self.value_cache[layer_idx]
